SwiftMiX 1.71
This version adds disc-caching and there is a re-write of the drag-drop move of multiple files and music-artist folders. You can right-click, select Edit Mode, hold either Ctrl or Shift down and choose a range of files and drag within the same list or to the other list - even while a song is playing. The playing song can be moved within the same list but not to the other list until it's stopped.