v1.83
In the RAD Studio C++ version, I did extensive change to the way media-faults are handled. It retries several times and tries the original file's path if the cached file fails - and if the song file has been deleted we force fade to the other player. And if the other player's songs have all played, we sequence to the next song on the current player. Getting all of that to work, harmoniously - is not so trivial!